pyproject.toml
[project]
name = "feedbax"
version = "0.1.2"
description = "Optimal feedback control + interventions in JAX."
# url = ""
authors = [
    {name = "MLL", email = "mll@mll.bio"}
]
license = {file = "LICENSE"}
readme = "README.md"
keywords = [
    'jax',
    'neural-networks',
    'optimal-control',
    'optimal-feedback-control',
    'pytorch',
    'biomechanics',
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "crc32c>=2.7.1",
    "diffrax>=0.7.0",
    "equinox>=0.11.12",
    "fastapi>=0.110.0",
    "jax>=0.5.2",
    "jaxtyping>=0.2.38",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "optax>=0.2.4",
    "plotly>=6.0.0",
    "polars>=1.24.0",
    "seaborn>=0.13.2",
    "tensorboardx>=2.6.2.2",
    "treescope>=0.1.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.29.0",
    "jax-cookbook",
    # From feedbax_experiments:
    "alembic>=1.15.1",
    "dill>=0.3.9",
    "kaleido==0.2.1",
    "pandas>=2.1.0",
    "pillow>=11.1.0",
    "pyexiv2>=2.15.3",
    "pyperclip>=1.9.0",
    "ruamel-yaml>=0.18.0",
    "rich>=13.9.4",
    "scikit-learn>=1.6.1",
]

[project.optional-dependencies]
dashboard = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
]
notebook = [
    "ipyfilechooser>=0.6.0",
    "ipywidgets>=8.1.5",
    "notebook>=7.4.1",
]
plotlyviewer = [
    "pyqt5>=5.15.11",
    "pyqtwebengine>=5.15.7",
]
mjx = [
    "mujoco>=3.0.0",
    "mujoco-mjx>=0.1.0",
]

[project.scripts]
# Entry points from feedbax_experiments
feedbax-run = "feedbax.bin.run:main"
feedbax-analysis = "feedbax.bin.analysis:main"
feedbax-train = "feedbax.bin.train:main"

[project.urls]
repository = "https://github.com/i-m-mll/feedbax"
documentation = "https://docs.lprt.ca/feedbax"

[tool.uv]
config-settings = { editable-mode = "compat" }

#! TODO: Replace at release
[tool.uv.sources.jax-cookbook]
path = "/Users/mll/Main/10 Projects/05 Utils/jax-cookbook"
editable = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.12.11",
]

[tool.hatch.build.targets.sdist]
packages = ["feedbax"]

[tool.hatch.build.targets.wheel]
packages = ["feedbax"]

[tool.pyright]
include = ["feedbax", "tests"]
exclude = ["**/__pycache__", ".venv", ".claude"]
typeCheckingMode = "standard"
verboseOutput = true
reportMissingImports = true
reportMissingTypeStubs = true
# Temporary:
# reportAttributeAccessIssue = false

[tool.black]
target-version = ['py310', 'py311', 'py312']
include = 'feedbax\/.*\.pyi?$'
line-length = 88


[tool.ruff]
cache-dir = "~/.cache/ruff"
target-version = "py312"
line-length = 100

[tool.ruff.lint]
# Equinox array annotations
ignore = ["F722"]
# Prevent removal of unused imports
unfixable = ["F401"]

[tool.ruff.lint.per-file-ignores]
# Scripts often need to run things before importing more things
"feedbax/bin/*.py" = ["E402"]

[tool.ruff.format]
indent-style = "space"
